tweak avatar corner radius for issue #21
The radius is actually fixed at 3 dip for all avatar sizes (similar to
what happens on the site at the moment). This looks ok to me at the
sizes I've viewed so far, including the 16dp 'assigned' avatar - the
rounding is more subtle.

+++ app/src/main/java/com/github/mobile/android/util/Image.java
@@ -168,7 +168,7 @@ public class Image {
      * @param radius
      * @return rounded corner bitmap
      */
-    public static Bitmap roundCornersAndOverlayOnWhite(final Bitmap source, final int radius) {
+    public static Bitmap roundCornersAndOverlayOnWhite(final Bitmap source, final float radius) {
         int width = source.getWidth();
         int height = source.getHeight();
